pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "account-seller-bot"
version = "0.1.0"
description = "Telegram bot for selling accounts"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
dependencies = [
    "python-telegram-bot==20.6",
    "pyotp",
    "cryptography"
]

[project.scripts]
account-seller-bot = "bot:main"

[project.optional-dependencies]
dev = [
    "flake8",
    "pytest",
]

